LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-dropping-controls-expected.txt

Testing IOSInlineMediaControls and dropping controls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


TimeControl is visible
TimeControl is not visible

AirplayButton is visible
AirplayButton is not visible

PiPButton is visible
PiPButton is not visible

SkipBackButton is visible
SkipBackButton is not visible

FullscreenButton is visible
FullscreenButton is not visible

PASS successfullyParsed is true

TEST COMPLETE
